Create package for timing and change implementation of the Timer class
